Refactored fragmentation model and added more control for neutral losses from glycans and immonium ions
